requirements.txt
fastapi
requests
uvicorn
python-dotenv